# openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /magic:
    post:
      tags:
        - magic-controller
      operationId: getRandomMagic
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Magic'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Magic'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /word:
    get:
      tags:
        - word-controller
      operationId: getRandomWord
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Word'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /quote:
    get:
      tags:
        - quote-controller
      operationId: getRandomQuote
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Quote'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
components:
  schemas:
    CustomErrorResponse:
      type: object
      properties:
        errorMsg:
          type: string
        status:
          type: integer
          format: int32
        errorCode:
          type: string
        timeStamp:
          type: string
          format: date-time
    Magic:
      required:
        - question
      type: object
      properties:
        id:
          type: integer
          format: int32
        question:
          type: string
        answer:
          type: string
    Word:
      type: object
      properties:
        id:
          type: integer
          format: int32
        word:
          type: string
        definition:
          type: string
    Quote:
      type: object
      properties:
        id:
          type: integer
          format: int32
        author:
          type: string
        quote:
          type: string